extracted credentials exceptions in separated file
