+++ backend/exceptions/credentials_exceptions.py
@@ -1,0 +1,14 @@
+from fastapi import HTTPException, status
+
+
+credentials_exception = HTTPException(
+    status_code=status.HTTP_401_UNAUTHORIZED,
+    detail="Incorrect username or password",
+    headers={"WWW-Authenticate": "Basic"},
+)
+
+authentication_scheme_exception = HTTPException(
+    status_code=status.HTTP_401_UNAUTHORIZED,
+    detail="Invalid authentication scheme",
+    headers={"WWW-Authenticate": "Basic"},
+)
