+++ backend/utils/tests/test_fastapi.py
@@ -1,7 +1,7 @@
 import pytest
 
 from ..fastapi import scan_platform, scan_rom
-from ..exceptions import RomsNotFoundException
+from ...exceptions.fs_exceptions import RomsNotFoundException
 from models import Platform, Rom
